Extract selectWorkspaceFolder to separate utility file
Move selectWorkspaceFolder from WorkspacePicker.tsx to lib/workspace.ts
to satisfy react-refresh/only-export-components rule without needing
eslint-disable.

This separates utility functions from component exports, which is cleaner
and allows Fast Refresh to work properly.
