+++ src/renderer/src/components/chat/ChatContainer.tsx
@@ -7,7 +7,8 @@ import { useCurrentThread, useThreadStream } from '@/lib/thread-context'
 import { MessageBubble } from './MessageBubble'
 import { ModelSwitcher } from './ModelSwitcher'
 import { Folder } from 'lucide-react'
-import { WorkspacePicker, selectWorkspaceFolder } from './WorkspacePicker'
+import { WorkspacePicker } from './WorkspacePicker'
+import { selectWorkspaceFolder } from '@/lib/workspace'
 import { ChatTodos } from './ChatTodos'
 import { ContextUsageIndicator } from './ContextUsageIndicator'
 import type { Message } from '@/types'
