+++ src/renderer/src/components/chat/WorkspacePicker.tsx
@@ -1,37 +1,10 @@
-/* eslint-disable react-refresh/only-export-components */
 import { useState, useEffect } from 'react'
 import { Folder, Check, ChevronDown } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
 import { useCurrentThread } from '@/lib/thread-context'
 import { cn } from '@/lib/utils'
-import type { FileInfo } from '@/types'
-
-export async function selectWorkspaceFolder(
-  currentThreadId: string | null,
-  setWorkspacePath: (path: string | null) => void,
-  setWorkspaceFiles: (files: FileInfo[]) => void,
-  setLoading: (loading: boolean) => void,
-  setOpen?: (open: boolean) => void
-): Promise<void> {
-  if (!currentThreadId) return
-  setLoading(true)
-  try {
-    const path = await window.api.workspace.select(currentThreadId)
-    if (path) {
-      setWorkspacePath(path)
-      const result = await window.api.workspace.loadFromDisk(currentThreadId)
-      if (result.success && result.files) {
-        setWorkspaceFiles(result.files)
-      }
-    }
-    if (setOpen) setOpen(false)
-  } catch (e) {
-    console.error('[WorkspacePicker] Select folder error:', e)
-  } finally {
-    setLoading(false)
-  }
-}
+import { selectWorkspaceFolder } from '@/lib/workspace'
 
 interface WorkspacePickerProps {
   threadId: string
